+++ src/ManyUI.jl
@@ -11,6 +11,7 @@ module ManyUI
 using Unicode
 using InlineStrings
 using Markdown
+using JuliaSyntaxHighlighting
 using DocStringExtensions
 import REPL
 
@@ -83,6 +84,7 @@ include("widgets/progresslist.jl")
 include("widgets/statusbar.jl")
 include("widgets/splitter.jl")
 include("widgets/markdownpane.jl")
+include("widgets/codeeditor.jl")
 include("widgets/dialog.jl")
 include("widgets/form.jl")
 include("precompile.jl")
@@ -203,6 +205,10 @@ export StatusBar, segments, status_layout
 # widgets/progresslist.jl
 export ProgressList, ProgressItem, n_items, set_progress!
 export pl_label_width, PL_GAP, PL_MIN_BAR
+
+# widgets/codeeditor.jl
+export CodeEditor, code_lines, highlight_julia, code_face_style
+export CODE_FACES
 
 # widgets/markdownpane.jl
 export MarkdownPane, set_source!, md_lines, md_inline